requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
azure-storage-blob==12.19.1
python-dotenv==1.0.1
python-jose==3.3.0
aiofiles==23.2.1
faster-whisper
torch
labelbox
pandas
openpyxl
ffmpeg-python
ctranslate2
openai
aiohttp